+++ reahl-doc/.reahlproject
@@ -46,8 +46,8 @@
         <class locator="reahl.doc.examples.tutorial.addressbook1.addressbook1:Address"/>
         <class locator="reahl.doc.examples.tutorial.pageflow2.pageflow2:Address"/>
         <class locator="reahl.doc.examples.tutorial.pageflow1.pageflow1:Address"/>
-        <class locator="reahl.doc.examples.tutorial.parameterised1bootstrap.parameterised1bootstrap:Address"/>
-        <class locator="reahl.doc.examples.tutorial.parameterised2bootstrap.parameterised2bootstrap:Address"/>
+        <class locator="reahl.doc.examples.tutorial.parameterised1.parameterised1:Address"/>
+        <class locator="reahl.doc.examples.tutorial.parameterised2.parameterised2:Address"/>
         <class locator="reahl.doc.examples.tutorial.sessionscopebootstrap.sessionscopebootstrap:User"/>
         <class locator="reahl.doc.examples.tutorial.sessionscopebootstrap.sessionscopebootstrap:LoginSession"/>
         <class locator="reahl.doc.examples.tutorial.access1bootstrap.access1bootstrap:AddressBook"/>
