+++ reahl-doc/reahl/doc/examples/tutorial/parameterised1/.reahlproject
@@ -19,9 +19,9 @@
   </deps>
 
   <persisted>
-    <class locator="reahl.doc.examples.tutorial.parameterised1bootstrap.parameterised1bootstrap:Address"/>
+    <class locator="reahl.doc.examples.tutorial.parameterised1.parameterised1:Address"/>
   </persisted>
   
-  <alias name="unit" command="setup -- -q pytest --addopts '-s --pyargs reahl.doc.examples.tutorial.parameterised1bootstrap.parameterised1bootstrap_dev'"/>
+  <alias name="unit" command="setup -- -q pytest --addopts '-s --pyargs reahl.doc.examples.tutorial.parameterised1.parameterised1_dev'"/>
   
 </project>
