+++ reahl-doc/reahl/doc/examples/tutorial/parameterised1/etc/web.config.py
@@ -1,6 +1,6 @@
 
 
 from __future__ import print_function, unicode_literals, absolute_import, division
-from reahl.doc.examples.tutorial.parameterised2bootstrap.parameterised2bootstrap import AddressBookUI
+from reahl.doc.examples.tutorial.parameterised1.parameterised1 import AddressBookUI
 
 web.site_root = AddressBookUI
