+++ reahl-doc/reahl/doc/examples/tutorial/parameterised1/parameterised1.py
@@ -1,0 +1,128 @@
+from __future__ import print_function, unicode_literals, absolute_import, division
+
+
+from reahl.web.fw import UserInterface, Widget, CannotCreate, UrlBoundView
+from reahl.web.layout import PageLayout
+from reahl.web.bootstrap.ui import HTML5Page, TextNode, Div, H, P, A
+from reahl.web.bootstrap.navbar import Navbar, ResponsiveLayout
+from reahl.web.bootstrap.navs import Nav
+from reahl.web.bootstrap.grid import Container, ColumnLayout, ColumnOptions, ResponsiveSize
+from reahl.web.bootstrap.forms import TextInput, Form, FormLayout, Button, ButtonLayout, FieldSet
+from reahl.component.modelinterface import exposed, Field, EmailField, Action, Event, IntegerField
+from reahl.sqlalchemysupport import Session, Base
+from sqlalchemy.orm.exc import NoResultFound
+from sqlalchemy import Column, Integer, UnicodeText
+
+
+class AddressBookPage(HTML5Page):
+    def __init__(self, view, bookmarks):
+        super(AddressBookPage, self).__init__(view)
+        self.use_layout(PageLayout(document_layout=Container()))
+        contents_layout = ColumnLayout(ColumnOptions('main', size=ResponsiveSize())).with_slots()
+        self.layout.contents.use_layout(contents_layout)
+
+        layout = ResponsiveLayout('md', colour_theme='inverse', bg_scheme='primary', toggle_button_alignment='right')
+        navbar = Navbar(view, css_id='my_nav').use_layout(layout)
+        navbar.layout.set_brand_text('Address book')
+        navbar.layout.add(Nav(view).with_bookmarks(bookmarks))
+        navbar.layout.add(TextNode(view, 'All your addresses in one place'))
+
+        self.layout.header.add_child(navbar)
+
+
+class EditView(UrlBoundView):
+    def assemble(self, address_id=None):
+        try:
+            address = Session.query(Address).filter_by(id=address_id).one()
+        except NoResultFound:
+            raise CannotCreate()
+
+        self.title = 'Edit %s' % address.name
+        self.set_slot('main', EditAddressForm.factory(address))
+
+
+class EditAddressForm(Form):
+    def __init__(self, view, address):
+        super(EditAddressForm, self).__init__(view, 'edit_form')
+
+        grouped_inputs = self.add_child(FieldSet(view, legend_text='Edit address'))
+        grouped_inputs.use_layout(FormLayout())
+        grouped_inputs.layout.add_input(TextInput(self, address.fields.name))
+        grouped_inputs.layout.add_input(TextInput(self, address.fields.email_address))
+
+        button = grouped_inputs.add_child(Button(self, address.events.update))
+        button.use_layout(ButtonLayout(style='primary'))
+
+
+class AddAddressForm(Form):
+    def __init__(self, view):
+        super(AddAddressForm, self).__init__(view, 'address_form')
+
+        new_address = Address()
+        grouped_inputs = self.add_child(FieldSet(view, legend_text='Add an address'))
+        grouped_inputs.use_layout(FormLayout())
+        grouped_inputs.layout.add_input(TextInput(self, new_address.fields.name))
+        grouped_inputs.layout.add_input(TextInput(self, new_address.fields.email_address))
+
+        button = grouped_inputs.add_child(Button(self, new_address.events.save))
+        button.use_layout(ButtonLayout(style='primary'))
+
+
+class AddressBookPanel(Div):
+    def __init__(self, view, address_book_ui):
+        super(AddressBookPanel, self).__init__(view)
+
+        self.add_child(H(view, 1, text='Addresses'))
+
+        for address in Session.query(Address).all():
+            self.add_child(AddressBox(view, address, address_book_ui))
+
+
+class AddressBox(Widget):
+    def __init__(self, view, address, address_book_ui):
+        super(AddressBox, self).__init__(view)
+        bookmark = address_book_ui.get_edit_bookmark(address=address, description='edit')
+        paragraph = self.add_child(P(view, text='%s: %s ' % (address.name, address.email_address)))
+        paragraph.add_child(A.from_bookmark(view, bookmark))
+
+
+class AddressBookUI(UserInterface):
+    def assemble(self):
+        show = self.define_view('/', title='Show')
+        add = self.define_view('/add', title='Add an address')
+        edit = self.define_view('/edit', view_class=EditView, address_id=IntegerField())
+
+        show.set_slot('main', AddressBookPanel.factory(self))
+        add.set_slot('main', AddAddressForm.factory())
+
+        bookmarks = [f.as_bookmark(self) for f in [show, add]]
+        self.define_page(AddressBookPage, bookmarks)
+
+        self.define_transition(Address.events.save, add, show)
+        self.define_transition(Address.events.update, edit, show)
+
+        self.edit = edit
+
+    def get_edit_bookmark(self, address, description=None):
+        return self.edit.as_bookmark(self, address_id=address.id, description=description)
+
+
+class Address(Base):
+    __tablename__ = 'parameterised1_address'
+
+    id            = Column(Integer, primary_key=True)
+    email_address = Column(UnicodeText)
+    name          = Column(UnicodeText)
+
+    @exposed
+    def fields(self, fields):
+        fields.name = Field(label='Name', required=True)
+        fields.email_address = EmailField(label='Email', required=True)
+
+    def save(self):
+        Session.add(self)
+
+    @exposed('save', 'update')
+    def events(self, events):
+        events.save = Event(label='Add', action=Action(self.save))
+        events.update = Event(label='Update')
