+++ reahl-doc/reahl/doc/examples/tutorial/parameterised2/etc/web.config.py
@@ -1,6 +1,6 @@
 
 
 from __future__ import print_function, unicode_literals, absolute_import, division
-from reahl.doc.examples.tutorial.parameterised1bootstrap.parameterised1bootstrap import AddressBookUI
+from reahl.doc.examples.tutorial.parameterised2.parameterised2 import AddressBookUI
 
 web.site_root = AddressBookUI
