+++ reahl-doc/reahl/doc/examples/tutorial/parameterised2/parameterised2_dev/test_parameterised2_2.py
@@ -1,13 +1,9 @@
-
-
 from __future__ import print_function, unicode_literals, absolute_import, division
+
 from reahl.tofu import Fixture, uses
 from reahl.tofu.pytestsupport import with_fixtures
-
 from reahl.webdev.tools import Browser, XPath
-
-from reahl.doc.examples.tutorial.parameterised2bootstrap.parameterised2bootstrap import AddressBookUI, Address
-
+from reahl.doc.examples.tutorial.parameterised2.parameterised2 import AddressBookUI, Address
 from reahl.web_dev.fixtures import WebFixture
